move some names around
